+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
++ 处理[issues 31](https://github.com/2fps/recorder/issues/31)。
 
 ## [1.0.2] 2020-02-15
 ### Fixed